Update README.md
The client, The Gaming Room,  wishes to create a game where teams will compete to guess the puzzle given certain stock drawings as clues. The game will consist of 4 rounds that each last one minute, and there may only be one instance of the game running at once, although there must be multiple teams made up of multiple players and team names may not repeat. The client wishes to develop a web-based version of the gaming app so that they can serve multiple platforms based on their current game that is only available for Android devices. The goal here was to make the game into a multi-platform game and this was accomplished by the recommendations I gave which fit the users' needs. In developing this documentation, I believe that I did a good job explaining my recommendations in ways that would make sense to anyone reading them, even if they have no background in computer science or software development. Using this plain terminology also helped me collect my thoughts and develop code that was concise and fit the user's needs. One thing that needs revising is the Domain model. I believe that I used too many computer-science-specific terms and it may be confusing for the client to understand.